test/backend/app/requirements.txt
# Python dependencies for backend application
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==1.4.41
pandas==1.4.3
scikit-learn==1.0.2
joblib==1.1.0
transformers==4.27.4
torch==1.12.1
spacy==3.4.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.2.0

